% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{all_less_than}
\alias{all_less_than}
\title{Find rows of a dataframe where all of the values
are below a threshold}
\usage{
all_less_than(df, threshold)
}
\arguments{
\item{df}{A (numeric) dataframe.}

\item{threshold}{A number.}
}
\value{
TRUE or FALSE.
}
\description{
Find rows of a dataframe where all of the values
are below a threshold
}

