% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{any_less_than}
\alias{any_less_than}
\title{Find rows of a dataframe where any of the values
is below a threshold}
\usage{
any_less_than(df, threshold)
}
\arguments{
\item{df}{A (numeric) dataframe.}

\item{threshold}{A number.}
}
\value{
TRUE or FALSE.
}
\description{
Find rows of a dataframe where any of the values
is below a threshold
}

