% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{apply_boundaries}
\alias{apply_boundaries}
\title{Filter data frame so all samples selected by `index` have read support
greater than the boundary threshold, or 0}
\usage{
apply_boundaries(df, boundaries, index)
}
\arguments{
\item{df}{A dataframe}

\item{boundaries}{Estimated boundaries from estimated_boundaries}

\item{index}{Columns of df to examine}
}
\value{
Filtered dataframe
}
\description{
Filter data frame so all samples selected by `index` have read support
greater than the boundary threshold, or 0
}

