% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{breakpoint_span}
\alias{breakpoint_span}
\title{Calculate distance between left and right breakpoint}
\usage{
breakpoint_span(df)
}
\arguments{
\item{df}{A dataframe}

\item{threshold}{Breakpoints must be at least this far apart, or on different chromosomes}
}
\value{
Vector of distances. NA returned for interchromosomal breaks
}
\description{
Calculate distance between left and right breakpoint
}

