% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{estimate_boundaries}
\alias{estimate_boundaries}
\title{Attempt to find the boundary between noisy, low-coverage SVs
and reliable high-coverage SVs}
\usage{
estimate_boundaries(df, index, exclude_above = 300)
}
\arguments{
\item{df}{A dataframe.}

\item{index}{The columns of the dataframe to operate on.}
}
\value{
TRUE or FALSE.
}
\description{
Attempt to find the boundary between noisy, low-coverage SVs
and reliable high-coverage SVs
}

