% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{one_greater_than}
\alias{one_greater_than}
\title{Find rows of a dataframe where exactly one of the values
is above a threshold}
\usage{
one_greater_than(df, threshold)
}
\arguments{
\item{df}{A (numeric) dataframe.}

\item{threshold}{A number.}
}
\value{
TRUE or FALSE.
}
\description{
Find rows of a dataframe where exactly one of the values
is above a threshold
}

