% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{variant_distance}
\alias{variant_distance}
\title{Retain SVs that span a distance greater than threshold}
\usage{
variant_distance(df, threshold)
}
\arguments{
\item{df}{A dataframe}

\item{threshold}{Breakpoints must be at least this far apart, or on different chromosomes}
}
\value{
A filtered dataframe
}
\description{
Retain SVs that span a distance greater than threshold
}

