% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{write_phylip}
\alias{write_phylip}
\title{Write data as binary sequence alignment in Phylip format}
\usage{
write_phylip(filename, data, threshold = 5)
}
\description{
Write data as binary sequence alignment in Phylip format
}

